Removing valgrind preference page (bug #353056)
This page is not longer needed because we will be able to set the
linuxtools PATH environment variable in Project Properties.

ChangeLog:
* src/org/eclipse/linuxtools/internal/valgrind/core/ValgrindCommand.java:
Removing dependency for ValgrindPreferencePage
* src/org/eclipse/linuxtools/internal/valgrind/core/ValgrindPlugin.java:
Removing dependency for ValgrindPreferencePage
* src/org/eclipse/linuxtools/internal/valgrind/core/ValgrindPreferencePage.java:
Removing this preference page. This page is not longer needed because we
will be able to set the linuxtools PATH environment variable in Project
Properties.
* src/org/eclipse/linuxtools/internal/valgrind/launch/ValgrindLaunchPlugin.java:
Removing dependency for ValgrindPreferencePage
* src/org/eclipse/linuxtools/internal/valgrind/core/ValgrindCommand.java:
Removing dependency for ValgrindPreferencePage
* src/org/eclipse/linuxtools/internal/valgrind/memcheck/tests/LocationPreferenceTest.java:
Removing dependency for ValgrindPreferencePage
* src/org/eclipse/linuxtools/internal/valgrind/tests/AbstractValgrindTest.java:
Removing dependency for ValgrindPreferencePage
* src/org/eclipse/linuxtools/internal/valgrind/tests/ValgrindStubCommand.java:
Removing dependency for ValgrindPreferencePage